# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "訂餐系統",
        "title": "Order Food API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "{{.Host}}",
    "basePath": "/api/",
    "paths": {
...
        "/manager/menu/item": {
            "post": {
                "description": "新增商品",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "新增商品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "商店",
                        "name": "shopID",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "商名",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "價格",
                        "name": "price",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "菜單",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/resp.Item"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/manager/menu/item/{id}": {
            "put": {
                "description": "修改商品",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "修改商品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "編號",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "品名",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "價格",
                        "name": "price",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "結果",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "刪除商品",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "刪除商品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "編號",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/manager/menu/item/{shopID}": {
            "get": {
                "description": "取得商品",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "取得商品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "商店編號",
                        "name": "shopID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "選單編號",
                        "name": "optionID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "菜單",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/resp.Item"
                            }
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/manager/menu/itemoption": {
            "post": {
                "description": "商品加入選單",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "商品加入選單",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "商品編號",
                        "name": "itemID",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "選單編號",
                        "name": "optionID",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "結果",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/resp.ItemOption"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/manager/menu/itemoption/{id}": {
            "delete": {
                "description": "刪除選單的商品",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "刪除選單的商品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/manager/menu/menu/{shop}": {
            "get": {
                "description": "取得菜單",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "取得菜單",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商店",
                        "name": "shop",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "菜單",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/resp.ShopMenu"
                            }
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/manager/menu/shopmenu/{shopID}": {
            "get": {
                "description": "取得菜單",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "取得菜單",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商店",
                        "name": "shopID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "菜單",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/resp.ShopMenu"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "resp.Item": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                },
                "Options": {
                    "type": "string"
                },
                "Price": {
                    "type": "integer"
                }
            }
        },
        "resp.ItemOption": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                },
                "ItemID": {
                    "type": "integer"
                },
                "OptionID": {
                    "type": "integer"
                }
            }
        },
        "resp.MenuOption": {
            "type": "object",
            "properties": {
                "Items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/resp.Item"
                    }
                },
                "Name": {
                    "type": "string"
                },
                "Option": {
                    "type": "object",
                    "$ref": "#/definitions/resp.Option"
                },
                "Selections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/resp.MenuSelection"
                    }
                }
            }
        },
        "resp.MenuSelection": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                },
                "Price": {
                    "type": "integer"
                }
            }
        },
        "resp.Option": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                },
                "SelectNum": {
                    "type": "integer"
                }
            }
        },
        "resp.Shop": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                }
            }
        },
        "resp.ShopMenu": {
            "type": "object",
            "properties": {
                "Options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/resp.MenuOption"
                    }
                },
                "Shop": {
                    "type": "object",
                    "$ref": "#/definitions/resp.Shop"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}